Refactor : 수정 서버빌드 확인
