--- src/main/java/com/sparta/devstar_be/board/Board.java
@@ -1,31 +1,0 @@
-package com.sparta.devstar_be.board;
-
-import jakarta.persistence.*;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Entity
-@Getter
-@NoArgsConstructor
-public class Board {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    @Column(nullable = false)
-    private String date;
-
-    @Column(nullable = false)
-    private String title;
-
-    @Column(nullable = false)
-    private String contents;
-
-    @Column
-    private String imageUrl;
-}
-
-    @ManyToOne
-    @JoinColumn(name = "user_id")
-    private User user;
-}
