+++ src/main/java/com/sparta/devstar_be/board/controller/BoardCommentsController.java
@@ -3,14 +3,10 @@ package com.sparta.devstar_be.board.controller;
 import com.sparta.devstar_be.board.dto.BoardCommentsDeleteReponseDto;
 import com.sparta.devstar_be.board.dto.BoardCommentsReponseDto;
 import com.sparta.devstar_be.board.dto.BoardCommentsRequestDto;
-import com.sparta.devstar_be.board.dto.BoardRequestDto;
 import com.sparta.devstar_be.board.service.BoardCommentsService;
-import com.sparta.devstar_be.security.UserDetailsImpl;
 import org.springframework.http.ResponseEntity;
-import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.bind.annotation.*;
-import org.springframework.web.servlet.function.EntityResponse;
 
 @RestController
 @RequestMapping("/api/starboards/{boardId}")
