+++ src/main/java/com/sparta/devstar_be/board/controller/BoardController.java
@@ -1,14 +1,10 @@
 package com.sparta.devstar_be.board.controller;
 
-import com.sparta.devstar_be.board.service.BoardService;
 import com.sparta.devstar_be.board.dto.BoardDeleteResponseDto;
 import com.sparta.devstar_be.board.dto.BoardRequestDto;
 import com.sparta.devstar_be.board.dto.BoardResponseDto;
-import com.sparta.devstar_be.enums.Major;
-import com.sparta.devstar_be.security.UserDetailsImpl;
-import lombok.RequiredArgsConstructor;
+import com.sparta.devstar_be.board.service.BoardService;
 import org.springframework.http.ResponseEntity;
-import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.web.bind.annotation.*;
 
 @RestController
