+++ src/main/java/com/sparta/devstar_be/like/LikeController.java
@@ -1,11 +1,10 @@
 package com.sparta.devstar_be.like;
 
-import com.sparta.devstar_be.security.UserDetailsImpl;
+import com.sparta.devstar_be.user.security.UserDetailsImpl;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
-import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.web.bind.annotation.*;
 
 @Slf4j
