+++ src/main/java/com/sparta/devstar_be/share/ShareController.java
@@ -1,13 +1,9 @@
 package com.sparta.devstar_be.share;
 
-import com.sparta.devstar_be.security.UserDetailsImpl;
 import jakarta.validation.Valid;
-import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
-import org.springframework.security.core.annotation.AuthenticationPrincipal;
-import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
