--- src/main/java/com/sparta/devstar_be/user/entity/UserRepository.java
@@ -1,6 +1,0 @@
-package com.sparta.devstar_be.user.entity;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface UserRepository extends JpaRepository<User,Long> {
-}
